fix uart and zifi in turbo mode x2 x4 (#113)
The UART code was taken and adapted from the latest Spectrum ZXUNO kernel. The behavior of rts(cts) was corrected so that it (and the RX state machine) remain inactive until the Z80 reads an incoming byte. A minor fix was made to the Zifi code to avoid duplicate incoming bytes caused by the new UART code. It is now more stable and can work with any Z80 clock signal.